added delete endpoints for step, feature, scenario. Closes #4

+++ src/main/java/com/tms/api/service/impl/StepServiceImpl.java
@@ -2,8 +2,10 @@ package com.tms.api.service.impl;
 
 import com.tms.api.data.dto.StepDto;
 import com.tms.api.data.entity.Step;
+import com.tms.api.data.repository.ScenarioRepository;
 import com.tms.api.data.repository.StepRepository;
 import com.tms.api.exception.AlreadyExistsException;
+import com.tms.api.exception.NotAllowedException;
 import com.tms.api.exception.ResourceNotFoundException;
 import com.tms.api.service.StepService;
 import com.tms.api.util.IdUtil;
@@ -20,7 +22,10 @@ import javax.annotation.PostConstruct;
 public class StepServiceImpl implements StepService {
 
     @Autowired
-    StepRepository repository;
+    private StepRepository repository;
+
+    @Autowired
+    private ScenarioRepository scenarioRepository;
 
     @Autowired
     private ModelMapper mapper;
@@ -61,9 +66,17 @@ public class StepServiceImpl implements StepService {
 
     @Override
     public void deleteById(String id) {
-        Step step = repository
-                .findByStepId(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Step id doesn't exists."));
-        repository.delete(step);
+        boolean isStepAssociatedWithScenario = scenarioRepository.findAll().stream()
+                .flatMap(scenario -> scenario.getSteps().stream())
+                .anyMatch(step -> step.getStepId().equals(id));
+        if (isStepAssociatedWithScenario) {
+            throw new NotAllowedException("Can't delete this step. Step is associated with scenario.");
+        } else {
+            Step step = repository.findByStepId(id).orElseThrow(() -> new ResourceNotFoundException("Step id doesn't exist."));
+            repository
+                    .findByStepId(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Step id doesn't exists."));
+            repository.delete(step);
+        }
     }
 }
